pyproject.toml
[project]
name = "pdf-extraction"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "pymupdf>=1.24.0",
    "Pillow>=10.3.0",
    "requests>=2.32.3",
    "tqdm>=4.66.4",
    "python-dotenv>=1.0.1",
    "pyarrow>=16.1.0",
    "pyqt6>=6.9.1",
    "shapely>=2.1.1",
    "pytesseract>=0.3.13",
]

[dependency-groups]
dev = [
    "ruff>=0.12.9",
]
